requirements.txt
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.2.1
pandas>=2.2.2
mysql-connector-python>=8.3.0
SQLAlchemy>=2.0.30
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
Flask-Login>=0.6.3
pytest>=8.2.0